Current work
This is roughly the stage I'm at currently, working on a few things at the same time, mainly trying to wrap my head around the process of taking the objects and performing the actions I need with them. I have some other stuff written down on paper but not implemented, I'm just putting this up so that I've commited something as I have to go to work now, feel free to email me though as I can reply on my phone when I get a break.